utests: seperate utests by features processing + update tools package + fix conftest, grpc server.
