+++ requirements/base.pip
@@ -16,5 +16,5 @@ pandas
 pydantic
 
 #
-pyconfr-2019-grpc-nlp-tools==0.1.1
-pyconfr-2019-grpc-nlp-protos==0.1.2
+pyconfr-2019-grpc-nlp-tools>=0.1.4
+pyconfr-2019-grpc-nlp-protos>=0.1.2
